+++ src/Commands/MigrationMakeCommand.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Nwidart\Modules\Commands;
+namespace Diegonella\Modules\Commands;
 
 use Illuminate\Support\Str;
-use Nwidart\Modules\Support\Config\GenerateConfigReader;
-use Nwidart\Modules\Support\Migrations\NameParser;
-use Nwidart\Modules\Support\Migrations\SchemaParser;
-use Nwidart\Modules\Support\Stub;
-use Nwidart\Modules\Traits\ModuleCommandTrait;
+use Diegonella\Modules\Support\Config\GenerateConfigReader;
+use Diegonella\Modules\Support\Migrations\NameParser;
+use Diegonella\Modules\Support\Migrations\SchemaParser;
+use Diegonella\Modules\Support\Stub;
+use Diegonella\Modules\Traits\ModuleCommandTrait;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;