+++ src/Lumen/Module.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Nwidart\Modules\Lumen;
+namespace Diegonella\Modules\Lumen;
 
 use Illuminate\Support\Str;
-use Nwidart\Modules\Module as BaseModule;
+use Diegonella\Modules\Module as BaseModule;
 
 class Module extends BaseModule
 {